+++ interface/src/lib/components/RSSIIndicator.svelte
@@ -18,17 +18,17 @@
 		<WiFi class={$$props.class || ''} />
 	{:else if rssi_dbm >= -75}
 		<div class="{$$props.class || ''} relative">
-			<WiFi class="absolute inset-0 h-full w-full opacity-10" />
+			<WiFi class="absolute inset-0 h-full w-full opacity-30" />
 			<WiFi2 class="absolute inset-0 h-full w-full" />
 		</div>
 	{:else if rssi_dbm >= -85}
 		<div class="{$$props.class || ''} relative">
-			<WiFi class="absolute inset-0 h-full w-full opacity-10" />
+			<WiFi class="absolute inset-0 h-full w-full opacity-30" />
 			<WiFi1 class="absolute inset-0 h-full w-full" />
 		</div>
 	{:else}
 		<div class="{$$props.class || ''} relative">
-			<WiFi class="absolute inset-0 h-full w-full opacity-10" />
+			<WiFi class="absolute inset-0 h-full w-full opacity-30" />
 			<WiFi0 class="absolute inset-0 h-full w-full" />
 		</div>
 	{/if}