+++ lib/framework/ESP32SvelteKit.cpp
@@ -42,6 +42,9 @@ ESP32SvelteKit::ESP32SvelteKit(AsyncWebServer *server) : _featureService(server)
 #if FT_ENABLED(FT_SLEEP)
                                                          _sleepService(server, &_securitySettingsService),
 #endif
+#if FT_ENABLED(FT_BATTERY)
+                                                         _batteryService(&_notificationEvents),
+#endif
                                                          _restartService(server, &_securitySettingsService),
                                                          _factoryResetService(server, &ESPFS, &_securitySettingsService),
                                                          _systemStatus(server, &_securitySettingsService)